+++ src/palette.rs
@@ -1,7 +1,11 @@
 // palette.rs   Color palette
 //
-// Copyright (c) 2019-2020  Douglas P Lau
+// Copyright (c) 2019-2023  Douglas P Lau
 //
+use crate::chan::{Ch8, Srgb, Straight};
+use crate::el::{Pix3, Pixel};
+use crate::gray::Gray8;
+use crate::raster::Raster;
 use crate::rgb::{Rgb, SRgb8};
 
 /// Color table for use with indexed `Raster`s.
@@ -138,6 +142,24 @@ impl Palette {
             }
         }
         Some(hist)
+    }
+
+    /// Make an indexed raster
+    pub fn make_indexed<S>(&mut self, raster: Raster<S>) -> Raster<Gray8>
+    where
+        S: Pixel<Chan = Ch8>,
+        <Pix3<Ch8, Rgb, Straight, Srgb> as Pixel>::Chan: From<S::Chan>,
+    {
+        let mut indexed = Raster::with_clear(raster.width(), raster.height());
+        for (src, dst) in raster.pixels().iter().zip(indexed.pixels_mut()) {
+            let clr = src.convert();
+            if let Some(e) = self.set_entry(clr) {
+                *dst = Gray8::new(e as u8);
+            } else {
+                // FIXME: handle full palette
+            }
+        }
+        indexed
     }
 }